Move pod sandbox recovery to podsandbox/ package
